AG-17010 Rename file to organization
