+++ packages/ag-charts-types/src/main.ts
@@ -73,7 +73,7 @@ export * from './series/standalone/flowProportionOptions';
 export * from './series/standalone/sankeyOptions';
 export * from './series/standalone/standaloneOptions';
 export * from './series/standalone/hierarchyOptions'; // Legacy
-export * from './series/standalone/organisationOptions';
+export * from './series/standalone/organizationOptions';
 export * from './series/standalone/pyramidOptions';
 export * from './series/standalone/sunburstOptions';
 export * from './series/standalone/treemapOptions';
